Revert "Add new widget for datalist"
This reverts commit 938ccff2fa702916c3e3aab9acda491e530e6d74.
